+++ Test/Piano_Violin.py
@@ -3,11 +3,6 @@ import cv2
 import threading
 import time
 
-from Player2 import play_violin_sound
-
-#from Pygame_playsound import play_violin_sound1
-
-
 scales = {'Major': ['40-4C.wav', '42-4D.wav', '44-4E.wav', '45-4F.wav', '47-4G.wav', '49-4A.wav', '51-4B.wav', '52-5C.wav'],
  'Minor': ['40-4C.wav', '42-4D.wav', '43-4D#.wav', '45-4F.wav', '47-4G.wav', '48-4G#.wav', '51-4B.wav', '52-5C.wav'],
   'Blues': ['40-4C.wav', '43-4D#.wav', '45-4F.wav', '46-4F#.wav', '47-4G.wav', '50-4A#.wav', '52-5C.wav', '55-5D#.wav'],
@@ -17,11 +12,17 @@ scales = {'Major': ['40-4C.wav', '42-4D.wav', '44-4E.wav', '45-4F.wav', '47-4G.w
       'Melodic Minor': ['40-4C.wav', '42-4D.wav', '43-4D#.wav', '45-4F.wav', '47-4G.wav', '49-4A.wav', '51-4B.wav', '52-5C.wav'],
        'Dorian Mode': ['40-4C.wav', '42-4D.wav', '43-4D#.wav', '45-4F.wav', '47-4G.wav', '49-4A.wav', '50-4A#.wav', '52-5C.wav'],
         'Mixolydian Mode': ['40-4C.wav', '42-4D.wav', '44-4E.wav', '45-4F.wav', '47-4G.wav', '49-4A.wav', '50-4A#.wav', '52-5C.wav'],
-         'Ahava Raba Mode': ['40-4C.wav', '41-4C#.wav', '44-4E.wav', '45-4F.wav', '47-4G.wav', '48-4G#.wav', '50-4A#.wav', '52-5C.wav']}
+         'Ahava Raba Mode': ['40-4C.wav', '41-4C#.wav', '44-4E.wav', '45-4F.wav', '47-4G.wav', '48-4G#.wav', '50-4A#.wav', '52-5C.wav'],
+         "BlackKeys": ["34-3F#.wav", "36-3G#.wav", "38-3A#.wav", "41-4C#.wav", "43-4D#.wav", "46-4F#.wav", "48-4G#.wav", "50-4A#.wav"],
+         "BlackPees": ["42-4D.wav", "44-4E.wav", "46-4F#.wav", "49-4A.wav", "50-4A#.wav", "51-4B.wav", "52-5C.wav", "56-5E.wav"]}
 
-# Settings for piano, want to make every half second so line 50 >= 0.5
+# Import your sound playing function
+from Sound_player import play_violin_sound
 
-
+# Replace with your sound playing function
+def play_sound_for_hand(hand_landmarks, hand_label):
+    # Implement your sound triggering logic here
+    pass
 
 mp_drawing = mp.solutions.drawing_utils
 mp_hands = mp.solutions.hands
@@ -29,28 +30,37 @@ mp_hands = mp.solutions.hands
 cap = cv2.VideoCapture(0)
 
 last_play_time = time.time()
-sound_thread = None  # Initialize sound_thread
 
+# Create a list to hold two sound threads and hand labels
+sound_threads = [None, None, None]
+hand_labels = ["Left Hand", "Right Hand", "Right Hand1"]
 
 insrument = "Piano"
 
-scale = "Major pentatonic"
+scale = "Blues"
+
+beats = 0.3
+
+h=480
+w=640
+
+# Coordinates to define regions for drawing borders
+border_lines = [(w // 8 * i, 0, w // 8 * i, h//2-20) for i in range(1, 8)]
+#make horizontal lines
+border_lines.append((0, h//2-20, w, h//2-20))
 
 
 def sound_thread_func(note):
     play_violin_sound(note, insrument)
 
-with mp_hands.Hands(min_detection_confidence=0.8, min_tracking_confidence=0.5) as hands:
-
-
+with mp_hands.Hands(min_detection_confidence=0.8, min_tracking_confidence=0.5, max_num_hands=3) as hands:
     while cap.isOpened():
-        
         ret, frame = cap.read()
 
         # BGR 2 RGB
         image = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
 
-        # Flip on horizontal
+        # Flip horizontally
         image = cv2.flip(image, 1)
 
         # Set flag
@@ -65,15 +75,18 @@ with mp_hands.Hands(min_detection_confidence=0.8, min_tracking_confidence=0.5) a
         # RGB 2 BGR
         image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
 
+        for x1, y1, x2, y2 in border_lines:
+            cv2.line(image, (x1, y1), (x2, y2), (255, 0, 0), 2)
+
         if results.multi_hand_landmarks:
             current_time = time.time()
-            if current_time - last_play_time >= 0.5:
-                for hand_landmarks in results.multi_hand_landmarks:
+            if current_time - last_play_time >= beats:
+                for hand_index, hand_landmarks in enumerate(results.multi_hand_landmarks):
                     for i, landmark in enumerate(hand_landmarks.landmark):
                         if i == 7:  # Landmark 8 (0-based index)
                             h, w, _ = image.shape
                             x, y = int(landmark.x * w), int(landmark.y * h)
-                            print(f"Landmark 8 (x, y): ({x}, {y})")
+                            print(f"{hand_labels[hand_index]} - Landmark 8 (x, y): ({x}, {y})")
 
                             if y < h/2:
 
@@ -134,9 +147,10 @@ with mp_hands.Hands(min_detection_confidence=0.8, min_tracking_confidence=0.5) a
                                     last_play_time = current_time
 
             # Draw hand landmarks
-            mp_drawing.draw_landmarks(image, hand_landmarks, mp_hands.HAND_CONNECTIONS,
-                                      mp_drawing.DrawingSpec(color=(121, 22, 76), thickness=2, circle_radius=4),
-                                      mp_drawing.DrawingSpec(color=(250, 44, 250), thickness=2, circle_radius=2))
+            for hand_index, hand_landmarks in enumerate(results.multi_hand_landmarks):
+                mp_drawing.draw_landmarks(image, hand_landmarks, mp_hands.HAND_CONNECTIONS,
+                                          mp_drawing.DrawingSpec(color=(121, 22, 76), thickness=2, circle_radius=4),
+                                          mp_drawing.DrawingSpec(color=(250, 44, 250), thickness=2, circle_radius=2))
 
         cv2.imshow('Hand Tracking', image)
 
@@ -145,7 +159,3 @@ with mp_hands.Hands(min_detection_confidence=0.8, min_tracking_confidence=0.5) a
 
 cap.release()
 cv2.destroyAllWindows()
-
-
-
-
