--- Test/plaer.py
@@ -1,3 +1,0 @@
-from Player2 import play_violin_sound as play
-
-play("Violin/44-3G.wav")
